Teensy cleanup

+++ neotrellis_monome_m0/MonomeSerialDevice.cpp
@@ -244,7 +244,7 @@ void MonomeSerialDevice::processSerial() {
                 if (i < deviceID.length()) {
                   Serial.write(deviceID[i]);
                 } else {
-                  Serial.write(0x00);
+                  Serial.write((uint8_t)0x00);
                 }
             }
             break;
